requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
google-generativeai==0.3.1
matplotlib==3.7.2
